Change to cast result to float rather than each individual component

+++ cpp2/ex01/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: kwang <kwang@student.42kl.edu.my>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/24 14:07:24 by kwang             #+#    #+#             */
-/*   Updated: 2022/03/23 15:30:56 by kwang            ###   ########.fr       */
+/*   Updated: 2022/03/24 17:29:21 by kwang            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ Fixed::Fixed(float const num)
 
 float Fixed::toFloat(void) const
 {
-	return (static_cast<float>(this->_value) / static_cast<float>((1 << this->_bits)));
+	return (static_cast<float>(this->_value / 1 << this->_bits));
 }
 
 int Fixed::toInt(void) const
